requirements.txt
boto3==1.36.26
pydantic-settings==2.8.0
SQLAlchemy==2.0.38